separate range dates to desktop and mobile file
